split up sockets and wifi
