+++ components/dis/CMakeLists.txt
@@ -4,5 +4,6 @@ idf_component_register(SRCS
                          "include"
                        REQUIRES
                          opendis
+                         sockets
                          wifi
 )
