+++ components/sockets/CMakeLists.txt
@@ -1,0 +1,8 @@
+idf_component_register(SRCS
+                         "src/sockets.cpp"
+                       INCLUDE_DIRS
+                         "include"
+                       REQUIRES
+                         constants
+                         lwip
+)
