+++ components/sockets/include/sockets.h
@@ -1,0 +1,4 @@
+#pragma once
+
+/// @brief Creates a UDP server
+int create_udp_server();
