+++ components/wifi/src/wifi.cpp
@@ -3,7 +3,6 @@
 #include <esp_log.h>
 #include <esp_wifi.h>
 #include <freertos/FreeRTOS.h>
-#include <lwip/sockets.h>
 #include <nvs_flash.h>
 
 #include "constants/general.h"
@@ -46,37 +45,6 @@ static void ip_event_handler(void* arg, esp_event_base_t event_base, int32_t eve
         retry_num = 0;
         xEventGroupSetBits(wifi_event_group, WIFI_SUCCESS);
     }
-}
-
-esp_err_t connect_udp_server() {
-    struct sockaddr_in serverInfo, clientInfo;
-    serverInfo.sin_family = AF_INET;
-    serverInfo.sin_addr.s_addr = INADDR_ANY;
-    serverInfo.sin_port = htons(46729);
-
-    char readBuffer[1024] = {0};
-
-    int sock = socket(AF_INET, SOCK_DGRAM, 0);
-    if (sock < 0) {
-        ESP_LOGE(TAG, "Failed to create a socket!");
-        return FAILURE;
-    }
-
-    if (bind(sock, (const struct sockaddr*)&serverInfo, sizeof(serverInfo)) != 0) {
-        ESP_LOGE(TAG, "Failed to connect to %s!", inet_ntoa(serverInfo.sin_addr.s_addr));
-        close(sock);
-        return FAILURE;
-    }
-
-    ESP_LOGI(TAG, "Connected to UDP server.");
-    while (1) {
-        socklen_t len = sizeof(clientInfo);
-        int n = recvfrom(sock, readBuffer, 1024, 0, (struct sockaddr*)&clientInfo, &len);
-        readBuffer[n] = '\0';
-        ESP_LOGI(TAG, "Received: %s\n", readBuffer);
-    }
-
-    return SUCCESS;
 }
 
 esp_err_t connect_wifi() {
@@ -145,12 +113,6 @@ int init_wifi() {
     // Establish a wifi connection
     if (connect_wifi() != SUCCESS) {
         ESP_LOGE(TAG, "Failed to connect to wifi!");
-        return FAILURE;
-    }
-
-    // Make a UDP server
-    if (connect_udp_server() != SUCCESS) {
-        ESP_LOGE(TAG, "Failed to start a TCP server!");
         return FAILURE;
     }
